# Memory bound merging for distributed aggregation in order