--- a/src/Core/SortDescription.h
+++ b/src/Core/SortDescription.h
@@ -125,6 +125,9 @@ public:
     bool hasPrefix(const SortDescription & prefix) const;
 };
 
+/// Returns a copy of lhs containing only the prefix of columns matching rhs's columns.
+SortDescription commonPrefix(const SortDescription & lhs, const SortDescription & rhs);
+
 /** Compile sort description for header_types.
   * Description is compiled only if compilation attempts to compile identical description is more than min_count_to_compile_sort_description.
   */