--- a/src/Planner/Planner.cpp
+++ b/src/Planner/Planner.cpp
@@ -495,7 +495,8 @@ void Planner::buildQueryPlanIfNeeded()
             settings.group_by_use_nulls,
             std::move(input_order_info),
             std::move(group_by_sort_description),
-            should_produce_results_in_order_of_bucket_number);
+            should_produce_results_in_order_of_bucket_number,
+            settings.enable_memory_bound_merging_of_aggregation_results);
         query_plan.addStep(std::move(aggregating_step));
 
         if (query_node.isGroupByWithRollup())