--- a/utils/check-style/codespell-ignore-words.list
+++ b/utils/check-style/codespell-ignore-words.list
@@ -23,3 +23,4 @@ hastable
 nam
 ubuntu
 toolchain
+vie