… src/Interpreters/InterpreterSelectQuery.cpp
… src/Processors/Merges/Algorithms/FinishAggregatingInOrderAlgorithm.cpp
… src/Processors/Merges/Algorithms/FinishAggregatingInOrderAlgorithm.h
… src/Processors/QueryPlan/AggregatingStep.cpp
… src/Processors/QueryPlan/AggregatingStep.h
… src/Processors/QueryPlan/IQueryPlanStep.h
… src/Processors/QueryPlan/ITransformingStep.h
… src/Processors/QueryPlan/MergingAggregatedStep.cpp
… src/Processors/QueryPlan/MergingAggregatedStep.h
… src/Processors/QueryPlan/ReadFromRemote.cpp
… src/Processors/QueryPlan/ReadFromRemote.h
… src/Processors/QueryPlan/UnionStep.cpp
… src/Processors/Transforms/AggregatingInOrderTransform.cpp
… src/Processors/Transforms/AggregatingTransform.h
… src/Processors/Transforms/MemoryBoundMerging.h
… src/Processors/Transforms/MergingAggregatedMemoryEfficientTransform.cpp
… src/Processors/Transforms/MergingAggregatedMemoryEfficientTransform.h
… src/Storages/MergeTree/MergeTreeDataSelectExecutor.cpp
… src/Storages/StorageDistributed.cpp
… tests/config/config.d/clusters.xml
… tests/integration/test_backward_compatibility/test_memory_bound_aggregation.py
… tests/performance/memory_bound_merging.xml
… tests/queries/0_stateless/02343_aggregation_pipeline.reference
… tests/queries/0_stateless/02404_memory_bound_merging.reference
… tests/queries/0_stateless/02404_memory_bound_merging.sql