@nickitat 2022-08-31
before:

``` sql
select avg(a) from remote('127.0.0.{1,2}', default, tt) group by a format Null settings optimize_aggregation_in_order = 1

0 rows in set. Elapsed: 35.488 sec. Processed 100.00 million rows, 800.00 MB (2.82 million rows/s., 22.54 MB/s.)

MemoryTracker: Peak memory usage (for query): 4.36 GiB.
```

after:

``` sql
select avg(a) from remote('127.0.0.{1,2}', default, tt) group by a format Null settings optimize_aggregation_in_order = 1

0 rows in set. Elapsed: 5.695 sec. Processed 100.00 million rows, 800.00 MB (17.56 million rows/s., 140.46 MB/s.)

MemoryTracker: Peak memory usage (for query): 286.06 MiB.
``